Move comparison lambdas into a separate subfolder
