+++ src/comparison/lambdas/countLambda.ts
