+++ src/comparison/lambdas/rerunComparisonLambda.integration.test.ts
@@ -5,20 +5,20 @@ process.env.BATCH_SIZE = "10"
 process.env.COMPARISON_LAMBDA_NAME = "DummyName"
 process.env.COMPARISON_BUCKET_NAME = comparisonBucketName
 
-import fs from "fs"
 import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3"
-import createS3Config from "./createS3Config"
-import MockS3 from "tests/helpers/MockS3"
-import MockDynamo from "tests/helpers/MockDynamo"
-import DynamoGateway from "./DynamoGateway/DynamoGateway"
-import createDynamoDbConfig from "./createDynamoDbConfig"
-import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
-import rerunComparisonLambda from "./rerunComparisonLambda"
-import compareLambda from "./compareLambda"
-import type { ComparisonLog } from "./Types"
-import { isError } from "./Types"
+import fs from "fs"
 import MockDate from "mockdate"
-import InvokeCompareLambda from "./InvokeCompareLambda"
+import MockDynamo from "tests/helpers/MockDynamo"
+import MockS3 from "tests/helpers/MockS3"
+import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
+import createDynamoDbConfig from "../createDynamoDbConfig"
+import createS3Config from "../createS3Config"
+import DynamoGateway from "../DynamoGateway/DynamoGateway"
+import InvokeCompareLambda from "../InvokeCompareLambda"
+import type { ComparisonLog } from "../Types"
+import { isError } from "../Types"
+import compareLambda from "./compareLambda"
+import rerunComparisonLambda from "./rerunComparisonLambda"
 
 const dynamoDbGatewayConfig = createDynamoDbConfig()
 const s3Config = createS3Config()
