+++ src/comparison/lambdas/rerunComparisonLambda.ts
@@ -1,9 +1,9 @@
 import logger from "src/lib/logging"
-import createDynamoDbConfig from "./createDynamoDbConfig"
-import DynamoGateway from "./DynamoGateway/DynamoGateway"
-import getFailedComparisonResults from "./getFailedComparisonResults"
-import InvokeCompareLambda from "./InvokeCompareLambda"
-import { isError } from "./Types"
+import createDynamoDbConfig from "../createDynamoDbConfig"
+import DynamoGateway from "../DynamoGateway/DynamoGateway"
+import getFailedComparisonResults from "../getFailedComparisonResults"
+import InvokeCompareLambda from "../InvokeCompareLambda"
+import { isError } from "../Types"
 
 const { BATCH_SIZE, COMPARISON_LAMBDA_NAME, COMPARISON_BUCKET_NAME } = process.env
 if (!BATCH_SIZE) {
